layer separation alongside utils "refactor"
