+++ src/domain/utils/errorHandler.ts
