+++ src/domain/utils/paths.ts
