+++ src/infrastructure/cardEditor.ts
@@ -3,7 +3,7 @@ import { AttackPage } from "./attackPage.js"
 import AttackType from "./attackType.js"
 import { Background } from "./background.js"
 import { Card } from "./card.js"
-import { CardDefinition } from "./cardDefinition.js"
+import { CardDefinition } from "../domain/cardDefinition.js"
 import Cost from "./cost.js"
 import Details from "./details.js"
 import { Overlay } from "./overlay.js"
